requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.4
aiosignal==1.3.1
async-timeout==5.0.1
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
contourpy==1.3.0
cycler==0.12.1
dataclasses-json==0.6.7
Deprecated==1.2.15
dune_client==1.7.7
fonttools==4.55.0
frozenlist==1.5.0
idna==3.10
importlib_resources==6.4.5
kiwisolver==1.4.7
marshmallow==3.23.1
matplotlib==3.9.2
multidict==6.1.0
mypy-extensions==1.0.0
ndjson==0.3.1
numpy==2.0.2
packaging==24.2
pandas==2.2.3
pillow==11.0.0
propcache==0.2.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
six==1.16.0
types-Deprecated==1.2.15.20241117
types-python-dateutil==2.9.0.20241003
types-PyYAML==6.0.12.20240917
types-requests==2.32.0.20241016
types-setuptools==75.5.0.20241119
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
wrapt==1.16.0
yarl==1.17.2
zipp==3.21.0
